data-insertion/requirements.txt
mysql-connector-python
prenoms